Added a formatdb application controller. This is primarily useful via the convenience functions provided in cogent/app/formatdb.py:
build_blast_db_from_seqs
build_blast_db_from_fasta_path

Which will let you build a temp blast database on the file, and return a tuple containing the name of the database and a list of the files to clean up. 

I've written a lot of simple functions to call formatdb lately, and noticed a lot of them sitting around in random places (untested) so I figured it was a good addition to cogent.




git-svn-id: https://pycogent.svn.sourceforge.net/svnroot/pycogent/trunk@445 4565cf14-4326-0410-8d81-eccf4396eebe