Several fixes: Changed original ChangedSequence to ChangedRnaSequence since it was RNA specific.  Added new ChangedSequence to have a cogent.core.sequence Sequence constructor as generic.
In RfamParser, seq_constructor parameter did nothing previously.  Now it actually does something (like use a passed in seq_constructor).  Default is ChangedRnaSequence as it was before, but now this will use whatever sequence constructor the user passes in.  Line 291 of RfamParser used to raise exception when using the strict parameter as False, which is incorrect behavior.  Now prints the Exception and specific error, so the continue on line 292 actually has a meaning, and the tests pass.

In MinimalRfamParser you can now pass in a seq_constructor unlike previously.  This is useful since a Minimal parser should not force sequences into a moltype without user being allowed to choose the sequence constructor.  Uses ChangedRnaSequence by default (as before).



git-svn-id: https://pycogent.svn.sourceforge.net/svnroot/pycogent/trunk@461 4565cf14-4326-0410-8d81-eccf4396eebe