Rewrote the app_path function. Rather than actually calling 'which' and working off the output, this now implements which.
This makes app_path more robust as we're not dependent on some string that gets returned by 'which', and that is different based on the version of 'which'. It also makes app_path more flexible because different environment variables can be checked, instead of only checking $PATH (see the env_variable paramter, which defaults to PATH).  


git-svn-id: https://pycogent.svn.sourceforge.net/svnroot/pycogent/trunk@674 4565cf14-4326-0410-8d81-eccf4396eebe